Updating popper to v1.12.9
