+++ Bootstrap/Bootstrap2/Twitter-Bootstrap-API/src/main/scala/net/liftmodules/fobotb/snippet/FoBo/Bootstrap.scala
@@ -4,7 +4,7 @@ import net.liftweb.util._
 import net.liftweb.common._
 import net.liftweb.http._
 import Helpers._
-import net.liftmodules.fobotb.lib.BootstrapSH
+import net.liftmodules.fobotb.lib._
 
 /**
   * ==Bootstrap v2.x Snippet==
