+++ Bootstrap/Bootstrap3/TwBs-Bootstrap3-API/src/main/scala/net/liftmodules/fobobs/snippet/FoBo/Bs3ScriptHelper.scala
@@ -5,7 +5,7 @@ import net.liftweb.common._
 import net.liftweb.http._
 import Helpers._
 import net.liftweb.http.js._
-import net.liftmodules.fobobs.lib.BootstrapSH
+import net.liftmodules.fobobs.lib._
 
 /**
   * ==Script Helper Snippet Bootstrap v3.x==
