+++ FoBo/FoBo/src/test/scala/net/liftmodules/fobo/FoBoSpec.scala
@@ -1188,44 +1188,44 @@ object FoBoSpec extends Specification {
   }
 
   //== Popper ==================================//
-  //Popper1123
-  "With fobo.Toolkit.init set to fobo.Toolkit.Popper1123 the ResourceServer" should {
+  //Popper1129
+  "With fobo.Toolkit.init set to fobo.Toolkit.Popper1129 the ResourceServer" should {
     "allow  fobo/popper.js" in {
-      allowToolkitInit(fobo.Toolkit.Popper1123, "fobo" :: "popper.js" :: Nil) must_== true
+      allowToolkitInit(fobo.Toolkit.Popper1129, "fobo" :: "popper.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper.js to fobo/popper/1.12.3/js/umd/popper.min.js" in {
-      rewriteToolkitInit(fobo.Toolkit.Popper1123, "fobo" :: "popper.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper.min.js")
+    "rewrit fobo/popper.js to fobo/popper/1.12.9/js/umd/popper.min.js" in {
+      rewriteToolkitInit(fobo.Toolkit.Popper1129, "fobo" :: "popper.js" :: Nil) must_==
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper.min.js")
     }
 
     "allow  fobo/popper-utils.js" in {
-      allowToolkitInit(fobo.Toolkit.Popper1123,
+      allowToolkitInit(fobo.Toolkit.Popper1129,
                        "fobo" :: "popper-utils.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper-utils.js to fobo/popper/1.12.3/js/umd/popper-utils.min.js" in {
-      rewriteToolkitInit(fobo.Toolkit.Popper1123,
+    "rewrit fobo/popper-utils.js to fobo/popper/1.12.9/js/umd/popper-utils.min.js" in {
+      rewriteToolkitInit(fobo.Toolkit.Popper1129,
                          "fobo" :: "popper-utils.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper-utils.min.js")
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper-utils.min.js")
     }
   }
 
-  "With fobo.Resource.init set to fobo.Resource.Popper1123 the ResourceServer" should {
+  "With fobo.Resource.init set to fobo.Resource.Popper1129 the ResourceServer" should {
     "allow  fobo/popper.js" in {
-      allowResource(fobo.Resource.Popper1123, "fobo" :: "popper.js" :: Nil) must_== true
+      allowResource(fobo.Resource.Popper1129, "fobo" :: "popper.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper.js to fobo/popper/1.12.3/js/umd/popper.min.js" in {
-      rewriteResource(fobo.Resource.Popper1123, "fobo" :: "popper.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper.min.js")
+    "rewrit fobo/popper.js to fobo/popper/1.12.9/js/umd/popper.min.js" in {
+      rewriteResource(fobo.Resource.Popper1129, "fobo" :: "popper.js" :: Nil) must_==
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper.min.js")
     }
 
     "allow  fobo/popper-utils.js" in {
-      allowResource(fobo.Resource.Popper1123,
+      allowResource(fobo.Resource.Popper1129,
                     "fobo" :: "popper-utils.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper-utils.js to fobo/popper/1.12.3/js/umd/popper-utils.min.js" in {
-      rewriteResource(fobo.Resource.Popper1123,
+    "rewrit fobo/popper-utils.js to fobo/popper/1.12.9/js/umd/popper-utils.min.js" in {
+      rewriteResource(fobo.Resource.Popper1129,
                       "fobo" :: "popper-utils.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper-utils.min.js")
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper-utils.min.js")
     }
   }
