+++ Popper/Popper-Res/src/test/scala/net/liftmodules/fobopop/fobopopresspec.scala
@@ -1,9 +1,7 @@
 package net.liftmodules
 
 import net.liftweb._
-import util.{Props}
 import http._
-import common._
 
 import org.specs2.mutable.Specification
 
@@ -12,26 +10,26 @@ object FoBoPopResSpec extends Specification {
 
   sequential
 
-  //Popper1123
-  "With FoBoPopRes.Resource.init set to FoBoPopRes.Resource.Popper1123 the ResourceServer" should {
+  //Popper1129
+  "With FoBoPopRes.Resource.init set to FoBoPopRes.Resource.Popper1129 the ResourceServer" should {
     "allow  fobo/popper.js" in {
-      allowResource(fobopopres.Resource.Popper1123,
+      allowResource(fobopopres.Resource.Popper1129,
                     "fobo" :: "popper.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper.js to fobo/popper/1.12.3/js/umd/popper.min.js" in {
-      rewriteResource(fobopopres.Resource.Popper1123,
+    "rewrit fobo/popper.js to fobo/popper/1.12.9/js/umd/popper.min.js" in {
+      rewriteResource(fobopopres.Resource.Popper1129,
                       "fobo" :: "popper.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper.min.js")
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper.min.js")
     }
 
     "allow  fobo/popper-utils.js" in {
-      allowResource(fobopopres.Resource.Popper1123,
+      allowResource(fobopopres.Resource.Popper1129,
                     "fobo" :: "popper-utils.js" :: Nil) must_== true
     }
-    "rewrit fobo/popper-utils.js to fobo/popper/1.12.3/js/umd/popper-utils.min.js" in {
-      rewriteResource(fobopopres.Resource.Popper1123,
+    "rewrit fobo/popper-utils.js to fobo/popper/1.12.9/js/umd/popper-utils.min.js" in {
+      rewriteResource(fobopopres.Resource.Popper1129,
                       "fobo" :: "popper-utils.js" :: Nil) must_==
-        List("fobo", "popper", "1.12.3", "js", "umd", "popper-utils.min.js")
+        List("fobo", "popper", "1.12.9", "js", "umd", "popper-utils.min.js")
     }
   }
